pyproject.toml
[project]
name = "lldm"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "diffusers>=0.33.1",
    "kaleido==0.2.1",
    "numpy>=2.2.5",
    "plotly>=6.0.1",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
    "wandb>=0.19.10",
]
[tool.ruff]
indent-width = 2
format.quote-style = "single"
ignore = ["E731"]
